remove Seo component, add Head component
