+++ src/components/common/Head/index.tsx
@@ -1,0 +1,69 @@
+import React from 'react'
+import { useStaticQuery, graphql } from 'gatsby'
+
+interface SEOProps {
+  description?: string
+  lang?: string
+  meta?: { [k: string]: string }[]
+  title: string
+}
+
+const Head = ({ description, lang, meta = [], title }: SEOProps) => {
+  const { site } = useStaticQuery(
+    graphql`
+      query {
+        site {
+          siteMetadata {
+            title
+            description
+            author
+          }
+        }
+      }
+    `
+  )
+
+  const metaDescription = description || site.siteMetadata.description
+
+  return (
+    <>
+      <title>{title}</title>
+      <meta name="description" content={metaDescription} />
+      <meta property="og:title" content={title} />
+      <meta
+        property="og:image"
+        content={require('/static/images/favicon.png')}
+      />
+      <meta property="og:description" content={metaDescription} />
+      <meta property="og:type" content="website" />
+      <meta name="twitter:card" content="summary" />
+      <meta name="twitter:creator" content={site.siteMetadata?.author || ''} />
+      <meta name="twitter:title" content={title} />
+      <meta name="twitter:description" content={metaDescription} />
+      {meta.map((props, i) => (
+        <meta key={i} {...props} />
+      ))}
+      <link rel="preconnect" href="https://fonts.gstatic.com" />
+      <link
+        href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&family=Anton&display=swap"
+        rel="stylesheet"
+      />
+      {/* eslint-disable-next-line react/no-danger */}
+      <div
+        dangerouslySetInnerHTML={{
+          __html: `
+          <script async defer src="https://www.googletagmanager.com/gtag/js?id=G-LNZQGS8EKR" />
+          <script>
+          window.dataLayer = window.dataLayer || [];
+          function gtag(){dataLayer.push(arguments);}
+          gtag('js', new Date());
+          gtag('config', 'G-LNZQGS8EKR');
+          </script>
+          `,
+        }}
+      />
+    </>
+  )
+}
+
+export default Head
